+++ Makefile-libpriv.am
@@ -30,6 +30,7 @@ librpmostreepriv_la_SOURCES = \
 	src/libpriv/rpmostree-refts.c \
 	src/libpriv/rpmostree-core.c \
 	src/libpriv/rpmostree-core.h \
+	src/libpriv/rpmostree-core-private.h \
 	src/libpriv/rpmostree-bwrap.c \
 	src/libpriv/rpmostree-bwrap.h \
 	src/libpriv/rpmostree-kernel.c \
